Deprecated R62 v1 and made ER62 the stable version
