+++ packages/alfa-rules/src/common/act/diagnostic.ts
@@ -1,9 +1,9 @@
-import { DistinguishingStyles as DistinguishingStylesExperimental } from "../../sia-er62/diagnostics";
+import { DistinguishingStyles } from "../../sia-r62/diagnostics";
 import { Languages } from "../../sia-r109/rule";
 import { LabelAndName } from "../../sia-r14/rule";
 import { RoleAndRequiredAttributes } from "../../sia-r16/rule";
 import { SameNames } from "../../sia-r56/rule";
-import { DistinguishingStyles } from "../../sia-r62/rule";
+import { DistinguishingStyles as DeprecatedDistinguishingStyles } from "../../sia-dr62/rule";
 import { DeprecatedElements } from "../../sia-r70/rule";
 import { WithDeclaration } from "../../sia-r75/rule";
 import { ClippingAncestors } from "../../sia-r83/rule";
@@ -35,7 +35,7 @@ export namespace Diagnostic {
   export const { isDistinguishingStyles } = DistinguishingStyles;
 
   export const { isDistinguishingStyles: isDistinguishingStylesExperimental } =
-    DistinguishingStylesExperimental;
+    DeprecatedDistinguishingStyles;
 
   export const { isLabelAndName } = LabelAndName;
