+++ packages/alfa-rules/src/deprecated.ts
@@ -1,0 +1,1 @@
+export { default as DR62 } from "./sia-dr62/rule";
