+++ packages/alfa-rules/src/experimental.ts
@@ -1,4 +1,3 @@
-export { default as ER62 } from "./sia-er62/rule";
 export { default as ER87 } from "./sia-er87/rule";
 export { default as R82 } from "./sia-r82/rule";
 export { default as R109 } from "./sia-r109/rule";
