+++ packages/alfa-rules/tsconfig.json
@@ -33,12 +33,12 @@
     "src/common/predicate/is-whitespace.ts",
     "src/common/predicate/is-wide-enough.ts",
     "src/common/predicate/reference-same-resource.ts",
+    "src/deprecated.ts",
     "src/experimental.ts",
     "src/index.ts",
     "src/rules.ts",
-    "src/sia-er62/rule.ts",
-    "src/sia-er62/diagnostics.ts",
-    "src/sia-er62/serialise.ts",
+    "src/sia-dr62/rule.ts",
+    "src/sia-dr62/serialise.ts",
     "src/sia-er65/rule.ts",
     "src/sia-er87/rule.ts",
     "src/sia-r1/rule.ts",
@@ -96,6 +96,7 @@
     "src/sia-r6/rule.ts",
     "src/sia-r60/rule.ts",
     "src/sia-r61/rule.ts",
+    "src/sia-r62/diagnostics.ts",
     "src/sia-r62/rule.ts",
     "src/sia-r62/serialise.ts",
     "src/sia-r63/rule.ts",
@@ -143,12 +144,8 @@
     "test/common/outcome.ts",
     "test/common/dom/get-colors.spec.tsx",
     "test/common/predicate/is-at-the-start.spec.tsx",
-    "test/sia-er62/rule.applicability.spec.tsx",
-    "test/sia-er62/rule.expectation.spec.tsx",
-    "test/sia-er62/rule.expectation.contrast.spec.tsx",
-    "test/sia-er62/rule.expectation.textdecoration.spec.tsx",
-    "test/sia-er62/serialise.spec.ts",
-    "test/sia-er62/common.ts",
+    "test/sia-dr62/rule.spec.tsx",
+    "test/sia-dr62/serialise.spec.ts",
     "test/sia-er65/rule.spec.tsx",
     "test/sia-er87/rule.spec.tsx",
     "test/sia-r1/rule.spec.tsx",
@@ -209,7 +206,11 @@
     "test/sia-r59/rule.spec.tsx",
     "test/sia-r60/rule.spec.tsx",
     "test/sia-r61/rule.spec.tsx",
-    "test/sia-r62/rule.spec.tsx",
+    "test/sia-r62/common.ts",
+    "test/sia-r62/rule.applicability.spec.tsx",
+    "test/sia-r62/rule.expectation.spec.tsx",
+    "test/sia-r62/rule.expectation.contrast.spec.tsx",
+    "test/sia-r62/rule.expectation.textdecoration.spec.tsx",
     "test/sia-r62/serialise.spec.ts",
     "test/sia-r63/rule.spec.tsx",
     "test/sia-r64/rule.spec.tsx",
